[core] changed cardinal class file content
